+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Extension can now work in browser (e.g. [vscode.dev](vscode.dev), [github.dev](github.dev)).
+
 ## [0.8.0] - 2022-07-20
 
 ### Added