+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added ability to choose multiple licenses. Files named in form `LICENSE-<SPDX_ID>`.
+- Added ability to choose multiple licenses. Files named in form `<filename>-<SPDX_ID><extension>`.
 
 ## [0.8.0] - 2022-07-20