+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added ability to choose multiple licenses. Files named in form `<filename>-<SPDX_ID><extension>`.
+- Extension can now work in browser (e.g. [vscode.dev](vscode.dev), [github.dev](github.dev)).
 
 ## [0.8.0] - 2022-07-20