+++ CHANGELOG.md
@@ -7,10 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.0] - 2022-09-11
+
 ### Added
 
 - Added ability to choose multiple licenses. Files named in form `<filename>-<SPDX_ID><extension>`.
 - Extension can now work in browser (e.g. [vscode.dev](vscode.dev), [github.dev](github.dev)).
+
+### Changed
+
+- Default license now always visible in quick pick menu (by is `always` meant when searching).
+- Make the `a` in the display name lowercase.
 
 ## [0.8.0] - 2022-07-20